requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.6.0
django-crontab==0.7.1
django-filter==24.3
django-heroku==0.3.1
django-picklefield==3.2
django-taggit==6.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
mysql-connector-python==9.0.0
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.7.0
